finals..
:)))

+++ files.h
@@ -4,14 +4,14 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <assert.h>
-#include<string.h>
+#include <string.h>
 
 // Struct datatypes declaration
 struct Files
 {
     char *input_source;
-	char *output_source;
-    size_t line_size;
+	char *output_source;	
+    int line_size;
 };
 
 struct Files *Files_Create(char *input_source, char *output_source, int line_size);
@@ -22,5 +22,6 @@ void Operate_Files_Actions(struct Files *m_File, char *word);
 void Close(FILE *fh);
 void Read_Std(char *line, size_t sz);
 void Write_Std(char *line);
+
 
 #endif
